requirements.txt
torch>=2.1
transformers>=4.48.0
accelerate>=0.30.0
datasets>=2.18.0
pillow>=10.0.0
tqdm>=4.66.0
